requirements.txt
django==3.1.5
djangorestframework==3.12.4
djoser==2.1.0
django-cors-headers==3.11.0
